+++ backend/integration_tests/src/client/multi_user.rs
@@ -1,6 +1,6 @@
 use crate::{generate_msgpack_query_call, generate_msgpack_update_call};
 use multi_user_canister::c2c_create_user;
-use user_canister::{bio, events, events_by_index, events_window, send_message_v2};
+use user_canister::{bio, delete_direct_chat, events, events_by_index, events_window, send_message_v2};
 
 // Queries
 generate_msgpack_query_call!(bio);
@@ -10,4 +10,5 @@ generate_msgpack_query_call!(events_window);
 
 // Updates
 generate_msgpack_update_call!(c2c_create_user);
+generate_msgpack_update_call!(delete_direct_chat);
 generate_msgpack_update_call!(send_message_v2);